requirements.txt
torch>=2.0
numpy>=1.23
pandas>=1.5
PyYAML>=6.0
